+++ packages/contracts/src/test/plugin/SharedPluginTest.sol
@@ -4,13 +4,13 @@ pragma solidity 0.8.17;
 
 import {Initializable} from "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
 
-import {IPermissionCondition} from "../../core/permission/IPermissionCondition.sol";
+import {PermissionConditionBase} from "../../core/permission/PermissionConditionBase.sol";
 import {PluginUUPSUpgradeable} from "../../core/plugin/PluginUUPSUpgradeable.sol";
 import {DaoUnauthorized} from "../../core/utils/auth.sol";
 import {IDAO} from "../../core/dao/IDAO.sol";
 
 /// @notice A test Plugin that manages permission to internal objects by associating their IDs with specific DAOs. Only the DAO for which the object was created has the permission to perform ID-gated actions on them.
-/// @dev This is realized by asking a `IPermissionCondition` that must be authorized in the DAO's permission manager.
+/// @dev This is realized by asking a `PermissionConditionBase` that must be authorized in the DAO's permission manager.
 contract TestSharedPlugin is PluginUUPSUpgradeable {
     bytes32 public constant ID_GATED_ACTION_PERMISSION_ID = keccak256("ID_GATED_ACTION_PERMISSION");
 
@@ -51,7 +51,7 @@ contract TestSharedPlugin is PluginUUPSUpgradeable {
     }
 
     /// @notice Executes something if the `id` parameter is authorized by the DAO associated through `ownedIds`.
-    ///         This is done by asking a `IPermissionCondition` that must be authorized in the DAO's permission manager via `grantWithCondition` and the `ID_GATED_ACTION_PERMISSION_ID`.
+    ///         This is done by asking a `PermissionConditionBase` that must be authorized in the DAO's permission manager via `grantWithCondition` and the `ID_GATED_ACTION_PERMISSION_ID`.
     /// @param _id The ID that is associated with a specific DAO
     function idGatedAction(uint256 _id) external sharedAuth(_id, ID_GATED_ACTION_PERMISSION_ID) {
         // do something
@@ -59,7 +59,7 @@ contract TestSharedPlugin is PluginUUPSUpgradeable {
 }
 
 /// @notice The condition associated with `TestSharedPlugin`
-contract TestIdGatingCondition is IPermissionCondition {
+contract TestIdGatingCondition is PermissionConditionBase {
     uint256 public allowedId;
 
     constructor(uint256 _id) {